+++ src/meta_extra.zig
@@ -15,14 +15,17 @@ pub inline fn isMarked(comptime T: type, comptime name: []const u8) bool {
     return false;
 }
 
-pub inline fn checkIterConstraints(
+pub inline fn expectIterSpecs(
     comptime I: type,
-    comptime name: []const u8,
+    comptime mark: []const u8,
     comptime Item: type,
-) bool {
-    if (!@hasField(I, "wrapped")) return false;
+) !void {
     const Wrapped = @FieldType(I, "wrapped");
-    return comptime I == Iter(Wrapped) and Wrapped.Item == Item and isMarked(Wrapped, name);
+        comptime {
+        if (I != Iter(Wrapped)) return error.NotAnIter;
+        if (Wrapped.Item != Item) return error.IncompatibleItemDecl;
+        if (!isMarked(Wrapped, mark)) return error.IncorrectIterWrapper;
+    }
 }
 
 pub inline fn expectImplIter(comptime T: type) !void {